requirements.txt
aiohttp==3.13.3
apscheduler==3.11.2
bleach==6.3.0
colorama==0.4.6
cryptography==46.0.4
fastapi==0.128.4
httpx==0.28.1
loguru==0.7.3
Markdown==3.10.1
matplotlib==3.10.8
openai==2.17.0
psutil==7.2.2
pydantic==2.12.5
PyJWT==2.8.0
PyJWT==2.11.0
pyquery==2.0.1
pytest==9.0.2
Requests==2.32.5
slowapi==0.1.9
SQLAlchemy==2.0.46
urllib3==2.6.3
uvicorn==0.40.0
